Change config to load multiple servers. Add default configuration.
To acommodate creation of SSL sockets support for multiple servers is
needed. To achieve this a list of servers in the config.json file is used.
The list can have as many servers as wanted, each on a different host/port
combination and optionally supporting SSL. A default config.json that
achieves the same as the original command line defaults has been included.